# src/infrastructure/openapi-internal.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AXPO Simulador de Ofertas - Internal API v1",
    "version": "0.2.0",
    "contact": {
      "name": "AXPO Development"
    },
    "description": "\n# Internal API\n\nSecure internal API for AXPO simulation management with role-based access control.\n\n## Features\n- RBAC (Admin/Agent/Commercial)\n- Simulation versioning and cloning\n- Base values management with versioning\n- Audit logging for all critical operations\n- Soft delete preservation\n\n## Authentication\n- **Bearer JWT token** (required)\n\n## Base URL\n- `/api/v1/internal`\n\n## Response Format\nAll responses follow a standard structure:\n```json\n{\n  \"success\": true/false,\n  \"data\": {...},\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human-readable message\",\n    \"details\": {...}\n  },\n  \"timestamp\": \"2026-03-11T10:00:00Z\"\n}\n```\n      "
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token for internal API access"
      },
      "tokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Public-Token",
        "description": "Public access token"
      },
      "pinAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PIN",
        "description": "4-digit PIN for public access"
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Internal API (requires Bearer JWT authentication)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/internal/config/email-templates": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get all email templates",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filter by template type"
          },
          {
            "in": "query",
            "name": "active",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          },
          {
            "in": "query",
            "name": "excludeType",
            "schema": {
              "type": "string"
            },
            "description": "Exclude templates of this type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of email templates"
          }
        }
      },
      "post": {
        "tags": [
          "Configuration"
        ],
        "summary": "Create a new email template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type",
                  "description",
                  "subject",
                  "htmlContent"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created email template"
          }
        }
      }
    },
...
    "/api/v1/internal/simulations/{id}/generate-pdf": {
      "post": {
        "summary": "Generate PDF from simulation HTML template",
        "tags": [
          "Simulations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Simulation ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "htmlContent"
                ],
                "properties": {
                  "htmlContent": {
                    "type": "string",
                    "description": "HTML content to convert to PDF"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF generated successfully",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}